removed what looks like dead code
The function finishDraw was removed. Afaics this function is dead
code.
The scene is now loaded before the simulation thread is started,
in case the -s parameter is specified.
The Graphics Core is now initialized before the simulation is started.
If the Graphics core is not started prior to the run simulation a segfault
in the laoding of the scene will happen.