load scene before starting the simulation thread
The scene is now loaded before the simulation thread is started,
in case the -s parameter is specified.

The Graphics Core is now initialized before the simulation is started.
If the Graphics core is not started prior to the run simulation a segfault
in the loading of the scene will happen.