correct box height and z pos in obstacle_generator
This allows boxes to be created only as large as needed, in contrast
to capsules which have to be twice as large larger as we only use
half of them. This means boxes have to be moved a bit further down on
an inclination such that not their middles, but their fronts
are in the correct height with respect to the inclination surface.

We still have to take another look at inclined obstacles, though.