obstacle_generator: simplify geometry calculation
A number of geometrical problems resulted from setting the boxes to
their actual height. It turns out that setting them at twice their
designated height and thus treat them like the capsules, where
only "half" of the geometry is used, simplifies a lot of things.
This is a results from the rotation always occuring at ground level,
which is where the origin of the obstacles is if they are twice
their specified height.