MotorData: motorMaxForce->maxEffort, maximumVelocity->maxSpeed
Member names in MotorData were changed to be consistent. Maximum speed
should be a scalar and does not have a direction as velocity has by
definition, and "motorMaxForce" was both redundant and unclear, as a
motor can exert a torque or a force depending on its type. The new names
are leaning on URDF's nomenclature.

Backwards compatibility is ensured by:
- keeping legacy getter and setter in SimMotor.h
- introducing a "legacynames" map in MotorData to allow correct parsing
  of old MARS scenes
Only old plugins directly accessing public members of MotorData should
need correction.