create nodes for all individual urdf objects
As the original implementation of the URDF-Loader in MARS caused
problems due to the combination of visual and collision objects
into nodes and moving the links' origins to their inertials, thereby
losing the origin of the link as defined in URDF, this revision instead
creates a node for every URDF object. This leads to a larger number of
nodes in the simulation but allows to retain all coordinate frames
defined in the URDF.

We could combine this SMURF entity factory with the one from the master
branch and introduce a flag that allows the user to specify (even in
SMURF) how to load a certain model.